requirements.txt
asgiref==3.3.4
Django==3.2.2
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
phonenumbers==8.12.23
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
